python_grpc/requirements.txt
autopep8==2.0.4
click==8.1.7
grpcio==1.60.1
grpcio-tools==1.60.1
humanfriendly==10.0
mypy==1.8.0
mypy-extensions==1.0.0
pause==0.3
protobuf==4.25.2
pycodestyle==2.11.1
types-protobuf==4.24.0.20240129
typing_extensions==4.9.0
